Refactor netcdf4 package.
