+++ datacube/drivers/netcdf/__init__.py
@@ -2,13 +2,3 @@
 #
 # Copyright (c) 2015-2026 ODC Contributors
 # SPDX-License-Identifier: Apache-2.0
-from . import writer as netcdf_writer
-from ._write import create_netcdf_storage_unit, write_dataset_to_netcdf
-from .writer import Variable
-
-__all__ = [
-    "Variable",
-    "create_netcdf_storage_unit",
-    "netcdf_writer",
-    "write_dataset_to_netcdf",
-]
